+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Changelog
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+Unreleased
+----------
+
+* Added `SysLanguageUidViewHelper`
+* Moved `LanguageViewHelper` under `Page` namespace
+
 1.2.1
 -----