+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Inject `LanguageService` into `AbstractRepository`
 * Improved `LanguageService`
 * Added `ArrayViewHelper`
 * Added `LocalizationUtility` (Draft)